requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
ldap3==2.9.1
python-dotenv==1.0.1
radius-eap-mschapv2-client==0.1.2
